enable adyen credit card min auth amount
Summary: enable adyen credit card min auth amount

Test Plan: bvt

Reviewers: fenglinyu

Subscribers: all

Differential Revision: https://code.silkcloud.info/D470

+++ payment/payment-clientproxy/src/main/java/com/junbo/payment/clientproxy/CurrencyServiceFacade.java
@@ -14,5 +14,5 @@ import com.junbo.langur.core.promise.Promise;
  */
 public interface CurrencyServiceFacade {
     Promise<Long> getMinAuthAmount(CurrencyId currencyId);
-    Promise<Integer> getNumberAfterDecimal(String currencyCode);
+    Promise<Long> getNumberAfterDecimal(String currencyCode);
 }
